Started splitting up upload/ edge for performance reasons
