+++ ansible/pistache.yml
@@ -1,0 +1,29 @@
+---
+
+- name: "Clone"
+  git:
+    repo: "https://github.com/oktal/pistache.git"
+    dest: "/usr/local/code/lib/pistache"
+    update: yes
+
+- name: "Set up build"
+  shell: "{{ item }}"
+  args:
+    chdir: "/usr/local/code/lib/pistache"
+  with_items:
+  - "git submodule update --init"
+  - "mkdir -p build"
+
+- name: "Build"
+  shell: "{{ item }}"
+  args:
+    chdir: "/usr/local/code/lib/pistache/build"
+  with_items:
+  - 'cmake -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Release ..'
+  - "make"
+
+- name: "Install"
+  become: yes
+  shell: "sudo make install"
+  args:
+    chdir: "/usr/local/code/lib/pistache/build"
