+++ include/edge/edge.h
@@ -7,10 +7,38 @@
 
 #include <cpprest/http_listener.h>
 
+#include <pistache/http.h>
+#include <pistache/router.h>
+#include <pistache/endpoint.h>
+
 using namespace web::http::experimental::listener;
 using namespace web::http;
 
+using namespace Pistache;
+
 namespace edge {
+    class FunctionEndpoint {
+    public:
+        explicit FunctionEndpoint(Address addr);
+
+        void init(int threadCount);
+
+        void start();
+
+        void shutdown();
+    private:
+        std::shared_ptr<Http::Endpoint> httpEndpoint;
+        Rest::Router router;
+
+        void setupRoutes();
+
+        void handleFunction(const Rest::Request &request, Http::ResponseWriter response);
+
+        void handleAsyncFunction(const Rest::Request &request, Http::ResponseWriter response);
+
+        message::FunctionCall buildCallFromRequest(const Rest::Request &request);
+    };
+
     class RestServer {
     public:
         RestServer();
@@ -18,8 +46,6 @@ namespace edge {
         void listen(const std::string &port);
 
         static void handleGet(const http_request &request);
-
-        static void handlePost(const http_request &request);
 
         static void handlePut(const http_request &request);
 
@@ -37,5 +63,4 @@ namespace edge {
 
     class InvalidPathException : public std::exception {
     };
-
 }
