+++ k8s/upload-pod.yml
@@ -1,0 +1,32 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: upload
+  namespace: faasm
+  labels:
+    app: faasm
+    role: upload
+spec:
+  containers:
+    - name: upload
+      image: faasm/upload
+      imagePullPolicy: Always
+      ports:
+        - containerPort: 8002
+      env:
+        - name: REDIS_HOST
+          value: redis
+        - name: REDIS_PORT
+          value: "6379"
+        - name: FUNC_ROOT
+          value: /usr/share/faasm
+        - name: LOG_LEVEL
+          value: "info"
+      volumeMounts:
+        - mountPath: "/usr/share/faasm"
+          name: func-vol
+    volumes:
+    - name: func-vol
+      persistentVolumeClaim:
+        claimName: func-storage-claim
+
