+++ src/edge/upload_server.cpp
@@ -1,7 +1,7 @@
 #include <edge/edge.h>
 
 int main() {
-    std::string port = "8001";
+    std::string port = "8002";
 
     util::initLogging();
